fix(derive): add revm-primitives as a dependency for kzg points
