+++ crates/derive/Cargo.toml
@@ -28,6 +28,7 @@ lru = "0.12.3"
 spin = { version = "0.9.8", features = ["mutex"] }
 c-kzg = { version = "1.0.0", default-features = false, optional = true }
 sha2 = { version = "0.10", default-features = false, optional = true }
+revm-primitives = { version = "3.1", default-features = false, optional = true }
 
 # `serde` feature dependencies
 serde = { version = "1.0.197", default-features = false, features = ["derive"], optional = true }
@@ -49,6 +50,7 @@ default = ["serde", "k256"]
 serde = ["dep:serde", "alloy-primitives/serde", "alloy-consensus/serde", "op-alloy-consensus/serde"]
 k256 = ["alloy-primitives/k256", "alloy-consensus/k256", "op-alloy-consensus/k256"]
 online = [
+  "dep:revm-primitives",
   "dep:c-kzg",
   "dep:sha2",
   "dep:alloy-provider",
@@ -56,4 +58,6 @@ online = [
   "dep:reqwest",
   "alloy-consensus/serde",
   "c-kzg/serde",
+  "revm-primitives/serde",
+  "revm-primitives/c-kzg",
 ]
