--- crates/derive/src/types/kzg/mod.rs
@@ -1,4 +1,0 @@
-//! Module for all things KZG related.
-
-mod points;
-pub use points::{G1Points, G2Points, G1_POINTS, G2_POINTS};
