--- crates/derive/src/types/kzg/points.rs
@@ -1,49 +1,0 @@
-//! KZG Trusted Setup Points.
-//!
-//! This code was adapted from [revm][revm].
-//!
-//! [revm]: https://github.com/bluealloy/revm/blob/main/crates/primitives/src/kzg/trusted_setup_points.rs
-
-use c_kzg::{BYTES_PER_G1_POINT, BYTES_PER_G2_POINT};
-
-/// Number of G1 Points.
-pub(crate) const NUM_G1_POINTS: usize = 4096;
-
-/// Number of G2 Points.
-pub(crate) const NUM_G2_POINTS: usize = 65;
-
-/// A newtype over list of G1 point from kzg trusted setup.
-#[derive(Debug, Clone, PartialEq)]
-#[repr(transparent)]
-pub struct G1Points(pub [[u8; BYTES_PER_G1_POINT]; NUM_G1_POINTS]);
-
-impl Default for G1Points {
-    fn default() -> Self {
-        Self([[0; BYTES_PER_G1_POINT]; NUM_G1_POINTS])
-    }
-}
-
-/// A newtype over list of G2 point from kzg trusted setup.
-#[derive(Debug, Clone, Eq, PartialEq)]
-#[repr(transparent)]
-pub struct G2Points(pub [[u8; BYTES_PER_G2_POINT]; NUM_G2_POINTS]);
-
-impl Default for G2Points {
-    fn default() -> Self {
-        Self([[0; BYTES_PER_G2_POINT]; NUM_G2_POINTS])
-    }
-}
-
-/// Default G1 points.
-pub const G1_POINTS: &G1Points = {
-    const BYTES: &[u8] = include_bytes!("./g1_points.bin");
-    assert!(BYTES.len() == core::mem::size_of::<G1Points>());
-    unsafe { &*BYTES.as_ptr().cast::<G1Points>() }
-};
-
-/// Default G2 points.
-pub const G2_POINTS: &G2Points = {
-    const BYTES: &[u8] = include_bytes!("./g2_points.bin");
-    assert!(BYTES.len() == core::mem::size_of::<G2Points>());
-    unsafe { &*BYTES.as_ptr().cast::<G2Points>() }
-};
